requirements.txt
flask
flask_sqlalchemy
python-dotenv
gunicorn
psycopg2-binary
